+++ cpsycle/ui/src/uilistbox.c
@@ -39,14 +39,14 @@ void ondestroy(ui_listbox* self, ui_component* sender)
 	signal_dispose(&self->signal_selchanged);
 }
 
-int ui_listbox_addstring(ui_listbox* listbox, const char* text)
+intptr_t ui_listbox_addstring(ui_listbox* listbox, const char* text)
 {
 	return SendMessage((HWND)listbox->component.hwnd, LB_ADDSTRING, 0, (LPARAM)text);
 }
 
-void ui_listbox_setstring(ui_listbox* self, const char* text, int index)
+void ui_listbox_setstring(ui_listbox* self, const char* text, intptr_t index)
 {
-	int sel;
+	intptr_t sel;
 
 	sel = ui_listbox_cursel(self);
 	SendMessage((HWND)self->component.hwnd, LB_DELETESTRING, (WPARAM)index, (LPARAM)text);
@@ -59,13 +59,12 @@ void ui_listbox_clear(ui_listbox* listbox)
 	SendMessage((HWND)listbox->component.hwnd, LB_RESETCONTENT, 0, (LPARAM)0);
 }
 
-void ui_listbox_setcursel(ui_listbox* listbox, int index)
+void ui_listbox_setcursel(ui_listbox* listbox, intptr_t index)
 {
-	SendMessage((HWND)listbox->component.hwnd, LB_SETCURSEL, (WPARAM)index, (LPARAM)0);
-	
+	SendMessage((HWND)listbox->component.hwnd, LB_SETCURSEL, (WPARAM)index, (LPARAM)0);	
 }
 
-int ui_listbox_cursel(ui_listbox* listbox)
+intptr_t ui_listbox_cursel(ui_listbox* listbox)
 {
 	return SendMessage((HWND)listbox->component.hwnd, LB_GETCURSEL, (WPARAM)0, (LPARAM)0);
 }
@@ -76,7 +75,7 @@ void ui_listbox_selitems(ui_listbox* listbox, int* items, int maxitems)
 		(LPARAM)items); 
 }
 
-int ui_listbox_selcount(ui_listbox* listbox)
+intptr_t ui_listbox_selcount(ui_listbox* listbox)
 {
 	return SendMessage((HWND)listbox->component.hwnd, LB_GETSELCOUNT, 
 		(WPARAM)0, (LPARAM)0); 
@@ -89,7 +88,7 @@ void oncommand(ui_listbox* self, ui_component* sender, WPARAM wParam,
         case LBN_SELCHANGE :
         {
             if (self->signal_selchanged.slots) {
-				int sel = SendMessage((HWND)self->component.hwnd, LB_GETCURSEL,
+				intptr_t sel = SendMessage((HWND)self->component.hwnd, LB_GETCURSEL,
 					(WPARAM)0, (LPARAM)0);
 				signal_emit(&self->signal_selchanged, self, 1, sel);
 			}
