+++ cpsycle/ui/src/uislider.h
@@ -16,6 +16,6 @@ void ui_slider_init(ui_slider*, ui_component* parent);
 void ui_slider_settext(ui_slider*, const char* text);
 void ui_slider_setrange(ui_slider*, int minrange, int maxrange);
 void ui_slider_setvalue(ui_slider*, int value);
-int ui_slider_value(ui_slider*);
+intptr_t ui_slider_value(ui_slider*);
 
 #endif
